feat: Unregister invalid configurations (#111)
When an invalid configuration file is given for a given site, the valid configuration that may have been previously registered for that site gets removed.
Same applies to the global custom configuration.
NB: A configuration file is considered invalid if it contains YAML syntax errors, unknown properties that do not exist in the equivalent Model objects, or invalid/missing values for some properties.