fix: i18n constraint error messages are not retrieved (#122)
Use hard-coded error messages for constraints error messages instead of using the i18n standard for JSR 303 validation.

In OSGi environment, the resource does not seem to be retrieved properly (when it's fine when running unit tests).
As those messages are only used to display error messages in the console, they don't actually need to be translated.