test(fix): Use a permission that exists in both 8.1.7 and 8.2 (#132)
Fix an issue with the nightly tests that are failing on a Jahia 8.1.7.

The root cause is that on Jahia 8.1.7.x, the permission view-full-wysiwyg-editor used to be present in the currentSite-access node for the Editor role, but is no longer present in Jahia 8.2.x.
The fix is to switch to the permission publish that, for both Jahia 8.1.7.x and 8.2.x), exists for the role Editor in chief and does not for the role Editor.